fix alias priority so alias_generators don't take priority
