+++ changes/1177-samuelcolvin.md
@@ -1,0 +1,1 @@
+`alias` takes priority over `alias_generator` on sub-classes
